Move dashboard drawer to drawers folder
